Moved animation settings to its own class
