+++ Assets/Scripts/Settings/UISettingsAnimation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb219ac831519664dbea3c160315d0ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
